+++ src/ts/components/closebutton.ts
@@ -24,13 +24,13 @@ export class CloseButton extends Button<CloseButtonConfig> {
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-closebutton',
       text: 'Close',
-    }, this.config);
+    } as CloseButtonConfig, this.config);
   }
 
   configure(player: PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
-    let config = <CloseButtonConfig>this.getConfig();
+    let config = this.getConfig();
 
     this.onClick.subscribe(() => {
       config.target.hide();